refactor(block): remove vestigial dedup layer (#1387, PR 2/3)
CAS StoreChunk already dedups physically by content hash, so the
explicit dedup layer is redundant: file-level dedup never hits (pending
donors never resolve to a durable target) and the block-level LRU's
AddRef path is Remote-gated, always falling back to StoreChunk for
pending CAS blocks. Remove both layers; every chunk now flows through
the idempotent StoreChunk with no change to manifests or BlockRefs.

- delete file-level dedup (dedup.go) and the flush pre-rollup hook
- delete block-level dedup LRU (dedup_lru.go) and the rollup consult
- drop the dedup_lru_size config knob
- remove the now-orphaned Syncer.coordinator field/setter and
  snapshotPendingBlockRefs helper
- move isObjectIDConflict into engine.go (still used by the rollup
  persist convergence path)
